refactor: 지원자 목록 관련 컴포넌트, 페이지 이름 변경
